refactor: files directory changed for proper folder structure
