+++ src/components/faq/faq.css
@@ -1,33 +1,3 @@
-/* .custom-content {
-  color: #f1f1f1;
-  font-size: 12px;
-  font-weight: 300;
-}
-
-.custom-content p {
-  font-size: 16px;
-}
-
-.custom-content ul {
-  margin: 0;
-  padding: 5px;
-  font-size: 12px;
-  line-height: 150%;
-}
-
-.custom-content li {
-  list-style: none;
-
-  margin-bottom: 4px;
-  font-size: 12px;
-  line-height: 150%;
-}
-
-#custom-content b {
-  color: #7feef0 !important;
-  font-weight: bold;
-} */
-
 .section {
   margin: 8px 0px 25px 0px;
 }
