+++ src/features/courts/hooks/useCourtDetailData.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getCourtDetail } from "../features/courts/api/courts.ts";
+import { getCourtDetail } from "../api/courts.ts";
 
 export const useCourtDetailData = ({ courtId }: { courtId: number }) => {
   return useQuery({
