+++ src/features/games/hooks/useFilterBox.tsx
