+++ src/features/games/hooks/useFilterBoxSettings.tsx
@@ -1,9 +1,12 @@
 import { useCallback, useMemo } from "react";
-import useDateSelectionStore from "../../store/useDateSelectionStore.ts";
-import useStatusSelectionStore from "../../store/useStatusSelectionStore.ts";
-import useTimeFieldStore from "../../store/useTimeStore.ts";
-import useGameFilterStore from "../../store/useGameFilterStore.ts";
-import { initialDateField, yearMonthToDate } from "../../constants/calender.ts";
+import useDateSelectionStore from "../../../store/useDateSelectionStore.ts";
+import useStatusSelectionStore from "../../../store/useStatusSelectionStore.ts";
+import useTimeFieldStore from "../../../store/useTimeStore.ts";
+import useGameFilterStore from "../../../store/useGameFilterStore.ts";
+import {
+  initialDateField,
+  yearMonthToDate,
+} from "../../../constants/calender.ts";
 
 export const useFilterBoxSettings = (
   setDisplayFilterBox: (arg: boolean) => void
