+++ src/features/inquiries/hooks/useInquiriesDataHook.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { privateInquiriesData } from "../api/support.ts";
+import { privateInquiriesData } from "../../../api/support.ts";
 
 export const useInquiriesDataHook = (page: number) => {
   return useQuery({
