+++ src/features/policies/hooks/useServiceTermsDataHook.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { policesData } from "../api/support.ts";
+import { policesData } from "../../../api/support.ts";
 
 export const useServiceTermsDataHook = () => {
   return useQuery({
